+++ R/data_write.R
@@ -15,6 +15,7 @@ data_write <- function(
   delimiter = ",",
   convert_factors = FALSE,
   save_labels = FALSE,
+  password = NULL,
   verbose = TRUE,
   ...
 ) {
@@ -52,30 +53,35 @@ data_write <- function(
       delimiter,
       convert_factors,
       save_labels,
+      password,
       verbose,
       ...
     )
   } else if (type == "rds") {
-    .write_rds(data, path, verbose, ...)
+    .write_rds(data, path, password, verbose, ...)
   } else if (type == "rda") {
-    .write_rda(data, path, verbose, ...)
+    .write_rda(data, path, password, verbose, ...)
   } else if (type == "parquet") {
-    .write_parquet(data, path, verbose, ...)
+    .write_parquet(data, path, password, verbose, ...)
   } else {
-    .write_haven(data, path, verbose, type, ...)
+    .write_haven(data, path, password, verbose, type, ...)
   }
 }
 
 
 # base R formats -----
 
-.write_rds <- function(data, path, verbose = TRUE, ...) {
+.write_rds <- function(data, path, password, verbose = TRUE, ...) {
+  # encrypt data
+  data <- .data_encryption(data, password)
   saveRDS(data, path, ...)
 }
 
-.write_rda <- function(data, path, verbose = TRUE, ...) {
+.write_rda <- function(data, path, password, verbose = TRUE, ...) {
+  # encrypt data
+  data <- .data_encryption(data, password)
   # save single data frame
-  if (is.data.frame(data)) {
+  if (is.data.frame(data) || is.raw(data)) {
     save(data, file = path, ...)
   } else {
     # save list of data frames
@@ -87,7 +93,15 @@ data_write <- function(
 
 # nanoparquet -----
 
-.write_parquet <- function(data, path, verbose = TRUE, ...) {
+.write_parquet <- function(data, path, password, verbose = TRUE, ...) {
+  # saving raw columns in data frames is not yet supported by parquet, thus,
+  # we cannot save encrypted data right now.
+  if (!is.null(password)) {
+    insight::format_error(
+      "Data encryption is not supported for parquet-files."
+    )
+  }
+
   # requires nanoparquet package
   insight::check_if_installed("nanoparquet")
 
@@ -105,9 +119,18 @@ data_write <- function(
   delimiter = ",",
   convert_factors = FALSE,
   save_labels = FALSE,
+  password = NULL,
   verbose = TRUE,
   ...
 ) {
+  # data encryption requires a data frame *with attributes* to be saved, which
+  # is not possible for text or raw formats - thus, no encryption here
+  if (!is.null(password)) {
+    insight::format_error(
+      "Data encryption is not supported for CSV or text files."
+    )
+  }
+
   # save labels
   if (save_labels && type == "csv") {
     data <- .save_labels_to_file(data, path, delimiter, verbose)
@@ -134,7 +157,22 @@ data_write <- function(
 
 # saving into haven format -----
 
-.write_haven <- function(data, path, verbose = TRUE, type = "spss", ...) {
+.write_haven <- function(
+  data,
+  path,
+  password,
+  verbose = TRUE,
+  type = "spss",
+  ...
+) {
+  # saving raw columns in data frames in not yet supported by haven, thus,
+  # we cannot save encrypted data right now.
+  if (!is.null(password)) {
+    insight::format_error(
+      "Data encryption is not supported for SPSS, SAS or Stata files."
+    )
+  }
+
   insight::check_if_installed("haven")
 
   # check if user provided "compress" argument for SPSS files,
@@ -340,4 +378,40 @@ data_write <- function(
 
   class(x) <- "data.frame"
   x
+}
+
+# data encryption ---------------------------------
+
+.data_encryption <- function(data, password = NULL) {
+  # check if data should be encrypted
+  if (!is.null(password)) {
+    .validate_password(password)
+    data <- .encrypt_data(data, password)
+    # tell user to remember the password
+    insight::format_warning(
+      "Remember your `password`, else you will not be able to decrypt the data again!"
+    )
+  }
+  data
+}
+
+.encrypt_data <- function(data, password = NULL) {
+  insight::check_if_installed("openssl", "for data encryption")
+  x <- serialize(data, NULL)
+  # it is important to remember the phrase! else, you cannot decrypt the data
+  passphrase <- charToRaw(password)
+  key <- openssl::sha256(passphrase)
+  # encrypt the data. we return the raw data here, which can be handled by
+  # rds/rda/rdata, and users can then also decrypt using openssl directly
+  # datawizard is not necessarily needed for decryption
+  openssl::aes_gcm_encrypt(x, key = key)
+}
+
+.validate_password <- function(password) {
+  # password needs to be a character string
+  if (!is.character(password) || length(password) != 1L || !nzchar(password)) {
+    insight::format_error(
+      "The password must be a single non-empty character string."
+    )
+  }
 }