+++ NEWS.md
@@ -9,6 +9,11 @@ CHANGES
 
 * `standardize()` now works on `fixest` estimations (#665).
 
+* `data_read()` and `data_write()` gain a `password` argument, to encrypt and
+  decrypt data files. This currently only works for R file formats (`.rda`,
+  `.rds`, and `.rdata`). Data encryption is based on the AES-GCM algorithm using
+  the `openssl::aes_gcm_encrypt()` function (#675).
+
 # datawizard 1.3.0
 
 BREAKING CHANGES